comment out direct links to WikiData properties
We wanted to use them like that in the past,
but this makes no sense, as WD is not RDF,
so we can not link to those properties (P****)
and classes (Q****) as if they were RDF entities.

Instead we have to look for properties in commonly used schemas
(like schema.org or FOAF), take similar properties from there
as out `supPropertyOf`'s, and then define a conversion from these
to the WD properties in the code in the conversion script in LOSH-tools.